Adjusted code to accomodate updates since it was last touched
Since this code was last used, I've updated many packages' versions
as well as Julia itself (from 1.8 -> 1.10). Here, I've updated the code
to rely on an old frozen version of Utilities instead of the live
version my other projects use. I've also tweaked a few spots in the code
to make it run correctly with these updated packages/code.

As of this commit, all of the behavior data funcs as well as the model
demonstration funcs for figure generation work and emit figures matching
what is in the original draft of the manuscript.